make semantic_parsing dataset reader directory
