+++ allennlp/data/dataset_readers/__init__.py
@@ -7,7 +7,6 @@ and produce datasets in the formats required by specific models.
 """
 
 # pylint: disable=line-too-long
-from allennlp.data.dataset_readers.atis import AtisDatasetReader
 from allennlp.data.dataset_readers.ccgbank import CcgBankDatasetReader
 from allennlp.data.dataset_readers.conll2003 import Conll2003DatasetReader
 from allennlp.data.dataset_readers.conll2000 import Conll2000DatasetReader
@@ -15,7 +14,6 @@ from allennlp.data.dataset_readers.ontonotes_ner import OntonotesNamedEntityReco
 from allennlp.data.dataset_readers.coreference_resolution import ConllCorefReader, WinobiasReader
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.dataset_readers.language_modeling import LanguageModelingReader
-from allennlp.data.dataset_readers.nlvr import NlvrDatasetReader
 from allennlp.data.dataset_readers.penn_tree_bank import PennTreeBankConstituencySpanDatasetReader
 from allennlp.data.dataset_readers.reading_comprehension import SquadReader, TriviaQaReader, QuACReader
 from allennlp.data.dataset_readers.semantic_role_labeling import SrlReader
@@ -25,5 +23,5 @@ from allennlp.data.dataset_readers.snli import SnliReader
 from allennlp.data.dataset_readers.universal_dependencies import UniversalDependenciesDatasetReader
 from allennlp.data.dataset_readers.stanford_sentiment_tree_bank import (
         StanfordSentimentTreeBankDatasetReader)
-from allennlp.data.dataset_readers.wikitables import WikiTablesDatasetReader
 from allennlp.data.dataset_readers.quora_paraphrase import QuoraParaphraseDatasetReader
+from allennlp.data.dataset_readers.semantic_parsing import WikiTablesDatasetReader, AtisDatasetReader, NlvrDatasetReader
