+++ allennlp/data/dataset_readers/semantic_parsing/__init__.py
@@ -1,0 +1,3 @@
+from allennlp.data.dataset_readers.semantic_parsing.atis import AtisDatasetReader
+from allennlp.data.dataset_readers.semantic_parsing.nlvr import NlvrDatasetReader
+from allennlp.data.dataset_readers.semantic_parsing.wikitables import WikiTablesDatasetReader
